requirements.txt
asgiref==3.5.0
autopep8==1.6.0
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-db-logger==0.1.12
django-request-logging==0.7.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.6.0
gunicorn==20.1.0
kafka==1.3.5
kafka-python==2.0.2
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing==3.7.4.3
tzdata==2021.5
virtualenv==20.13.1
waitress==2.1.0
